requirements.txt
selenium
beautifulsoup4
pdfplumber
google-generativeai